Reorganise search param code
